move legacy Icon
