+++ src/components/ButtonText/ButtonText.js
@@ -21,9 +21,9 @@ import * as React from 'react';
 import { Platform, TouchableOpacity, TouchableNativeFeedback } from 'react-native';
 import styled from 'styled-components/native';
 import { MediumText } from 'components/legacy/Typography';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Spinner from 'components/Spinner';
-import type { IconProps } from 'components/Icon';
+import type { IconProps } from 'components/legacy/Icon';
 import { fontSizes } from 'utils/variables';
 import { themedColors } from 'utils/themes';
