+++ src/components/CollapsibleSection/CollapsibleSection.js
@@ -23,7 +23,7 @@ import Collapsible from 'react-native-collapsible';
 
 // components
 import { MediumText, TextLink } from 'components/legacy/Typography';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 
 // utils
 import { fontSizes, fontStyles, spacing } from 'utils/variables';
