+++ src/components/ContactDetailsModal/ContactDetailsModal.js
@@ -32,7 +32,7 @@ import t from 'translations/translate';
 import { BaseText } from 'components/legacy/Typography';
 import AddressScanner from 'components/QRCodeScanner/AddressScanner';
 import Button from 'components/legacy/Button';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Image from 'components/Image';
 import Modal from 'components/Modal';
 import ModalBox from 'components/ModalBox';
