+++ src/components/EventDetails/EventDetails.js
@@ -31,7 +31,7 @@ import t from 'translations/translate';
 
 // components
 import { BaseText, MediumText } from 'components/legacy/Typography';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Image from 'components/Image';
 import TankAssetBalance from 'components/TankAssetBalance';
 import ReceiveModal from 'screens/Asset/ReceiveModal';
