+++ src/components/ListItem/SettingsItemCarded.js
@@ -21,7 +21,7 @@ import * as React from 'react';
 import styled, { withTheme } from 'styled-components/native';
 
 import ShadowedCard from 'components/ShadowedCard';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Image from 'components/Image';
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import Spinner from 'components/Spinner';
