+++ src/components/Switcher/SimpleSwitcher.js
@@ -22,7 +22,7 @@ import * as React from 'react';
 import { TouchableOpacity } from 'react-native';
 import styled from 'styled-components/native';
 
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import { BaseText } from 'components/legacy/Typography';
 import Spinner from 'components/Spinner';
