+++ src/components/Title/TitleWithIcon.js
@@ -21,7 +21,7 @@
 import * as React from 'react';
 import styled from 'styled-components/native';
 import { MediumText } from 'components/legacy/Typography';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import { fontStyles } from 'utils/variables';
 import { themedColors } from 'utils/themes';
