+++ src/components/legacy/Icon/package.json
