+++ src/components/legacy/TextInput/TextInput.js
@@ -30,7 +30,7 @@ import {
 import IconButton from 'components/IconButton';
 import { BaseText, MediumText } from 'components/legacy/Typography';
 import Spinner from 'components/Spinner';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Image from 'components/Image';
 import Button from 'components/legacy/Button';
 import Input from 'components/Input';
