+++ src/screens/Permissions/Permissions.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 // components
 import { Paragraph, MediumText, BaseText } from 'components/legacy/Typography';
 import Checkbox from 'components/Checkbox';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import CollapsibleListItem from 'components/ListItem/CollapsibleListItem';
 
 // utils
