+++ src/screens/PillarNetwork/PillarNetworkIntro.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import { ScrollWrapper } from 'components/legacy/Layout';
 import { BoldText, MediumText } from 'components/legacy/Typography';
-import Icon from 'components/Icon';
+import Icon from 'components/legacy/Icon';
 import Image from 'components/Image';
 import Button from 'components/legacy/Button';
 import { LabelBadge } from 'components/LabelBadge';
